Completed project The Race Day
Its an project about the runner whose age are under, equal & above 18 & whether they are early or late registered for the race. Using control-flow entering different combinations of values for the two variables ( runnerAge & statusForRegistration ) that I  created and run my code several times to verify that the correct statements are printing to the console!